Update Asset Client UI to use Pike Agents.
Update the asset client interface to use Pike Agents and organizations
instead of TnT agents. Because agent creation requires an associated
organization, the sign up form is now the 'get started' form that
creates both an organization and an admin user. Once Schema Records
are added, reporter functionality can be added back in as agent
metadata.

Signed-off-by: Davey Newhall <newhall@bitwise.io>